requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.5.18.1
et-xmlfile==1.1.0
flake8==4.0.1
mccabe==0.6.1
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
pycodestyle==2.8.0
pyflakes==2.4.0
PySimpleGUI==4.60.1
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2